Challenge: Add a second button to the app labeled "Count Up" that appears just below the Roll button. When tapped, the Count Up button should get the current value of the result text view, add 1 to it, and update the text view. Make sure you handle these edge cases:
If the result text view does not yet contain a number (that is, if the text view still has the default "Hello World" string), set the result text to 1.
If the number is already 6, do nothing.
Tip: The string in TextView views is an instance of class CharSequence. To test its value, you need to convert it to a string:

resultText.text.toString()

Coding challenge done solution code